Refactored code into apps
