+++ backend excercise/quiz/admin.py
@@ -1,4 +1,8 @@
 from django.contrib import admin
+
+from participation.models import Task, UserAnswer, UserQuiz, UserSubjectScore
+from quizzes.models import Question, Quiz, Subject
+from users.models import User
 from .models import *
 
 admin.site.register(User)
