+++ backend excercise/quizzes/serializers.py
@@ -1,0 +1,23 @@
+from rest_framework import serializers
+from .models import *
+
+
+
+class SubjectSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Subject
+        fields = '__all__'
+        lookup_field = 'Subject_ID'
+
+class QuizSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Quiz
+        fields = '__all__'
+        lookup_field = 'Quiz_ID'
+
+class QuestionSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Question
+        fields = '__all__'
+        lookup_field = 'Question_ID'
+
