+++ backend excercise/quizzes/views.py
@@ -1,0 +1,25 @@
+from django.shortcuts import render
+from rest_framework import viewsets, permissions
+
+from quizzes.models import Question, Quiz, Subject
+from quizzes.serializers import QuestionSerializer, QuizSerializer, SubjectSerializer
+
+# Create your views here.
+
+class SubjectViewSet(viewsets.ModelViewSet):
+    queryset = Subject.objects.all()
+    serializer_class = SubjectSerializer
+    permission_classes = [permissions.IsAuthenticated]
+    lookup_field = 'Subject_ID'
+
+class QuizViewSet(viewsets.ModelViewSet):
+    queryset = Quiz.objects.all()
+    serializer_class = QuizSerializer
+    permission_classes = [permissions.IsAuthenticated]
+    lookup_field = 'Quiz_ID'
+
+class QuestionViewSet(viewsets.ModelViewSet):
+    queryset = Question.objects.all()
+    serializer_class = QuestionSerializer
+    permission_classes = [permissions.IsAuthenticated]
+    lookup_field = 'Question_ID'
