+++ backend excercise/users/models.py
@@ -1,0 +1,41 @@
+import uuid
+from django.db import models
+from django.contrib.auth.models import AbstractBaseUser, PermissionsMixin, BaseUserManager
+# Create your models here.
+
+
+class UserManager(BaseUserManager):
+    def create_user(self, email, name, password=None, **extra_fields):
+        if not email:
+            raise ValueError('Email required')
+        email = self.normalize_email(email)
+        user = self.model(email=email, name=name, **extra_fields)
+        user.set_password(password)
+        user.save()
+        return user
+
+    def create_superuser(self, email, name, password=None, **extra_fields):
+        extra_fields.setdefault('is_staff', True)
+        extra_fields.setdefault('is_superuser', True)
+        return self.create_user(email, name, password, **extra_fields)
+    
+class User(AbstractBaseUser, PermissionsMixin):
+    User_ID = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
+    role = models.CharField(max_length=50)
+    name = models.CharField(max_length=50)
+    email = models.EmailField(max_length=255, unique=True)
+    is_active = models.BooleanField(default=True)
+    is_staff = models.BooleanField(default=False)
+
+    USERNAME_FIELD = 'email'
+    REQUIRED_FIELDS = ['name']
+
+    objects = UserManager()
+
+    # def save(self, *args, **kwargs):
+    #     if self.password and not self.Password_Hash:
+    #         self.set_password(self.password)
+    #     super().save(*args, **kwargs)
+    @property
+    def id(self):
+        return self.User_ID
